notices: move conditional logic inside components to prevent remounts
