+++ ui/src/components/UpdateNotice.tsx
@@ -1,11 +1,44 @@
-import React from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import AsteriskIcon from '@/components/icons/Asterisk16Icon';
 import { isTalk } from '@/logic/utils';
+import useKilnState, { usePike } from '@/state/kiln';
 
 export default function UpdateNotice() {
   const appName = isTalk ? 'Talk' : 'Groups';
-  function onClick() {
+  const pike = usePike(appName.toLocaleLowerCase());
+  const [baseHash, setBaseHash] = useState('');
+  const [needsUpdate, setNeedsUpdate] = useState(false);
+
+  const fetchPikes = useCallback(async () => {
+    try {
+      await useKilnState.getState().fetchPikes();
+    } catch (e) {
+      console.error(e);
+    }
+  }, []);
+
+  useEffect(() => {
+    if (pike && baseHash === '') {
+      setBaseHash(pike.hash);
+    }
+  }, [pike, baseHash]);
+
+  useEffect(() => {
+    setInterval(fetchPikes, 1000 * 60 * 5);
+  }, [fetchPikes]);
+
+  useEffect(() => {
+    if (pike && pike.hash !== baseHash && baseHash !== '' && !needsUpdate) {
+      setNeedsUpdate(true);
+    }
+  }, [pike, baseHash, needsUpdate]);
+
+  const onClick = useCallback(() => {
     window.location.reload();
+  }, []);
+
+  if (!needsUpdate) {
+    return null;
   }
 
   return (
